synthax+chore: Configure prettier/solhint for Solidity files
VS Code configured to auto-save using prettier as the default upadter.

https://github.com/prettier-solidity/prettier-plugin-solidity

+++ contracts/PozitionManager.sol
@@ -25,64 +25,60 @@ contract PozitionManager is ReentrancyGuard {
     /// Storage Variables ///
 
     /**
-    * @dev Synthetix's `AddressResolver` to grab Synths and Market contracts.
-    */
+     * @dev Synthetix's `AddressResolver` to grab Synths and Market contracts.
+     */
     IAddressResolver private addressResolver;
 
     /**
-    * @dev Margin token used for Synthetix Future positions. At the moment only sUSD is allowed.
-    */
+     * @dev Margin token used for Synthetix Future positions. At the moment only sUSD is allowed.
+     */
     IERC20 private marginToken;
 
     /**
-    * @dev Factory to mint 1/1 NFTs to represent an open position on Synthetix Futures.
-    */
+     * @dev Factory to mint 1/1 NFTs to represent an open position on Synthetix Futures.
+     */
     PozitionFactory private positionFactory;
 
     /**
-    * @dev A hardcoded map of supported markets.
-    */
-    mapping (bytes32 => IFuturesMarket) private supportedMarkets;
+     * @dev A hardcoded map of supported markets.
+     */
+    mapping(bytes32 => IFuturesMarket) private supportedMarkets;
 
     /// Events ///
 
     /**
-    * @dev Emitted with `amount` sUSD tokens are withdrawn from the margin vault by the `withdrawer`.
-    */
+     * @dev Emitted with `amount` sUSD tokens are withdrawn from the margin vault by the `withdrawer`.
+     */
     event Withdraw(address withdrawer, address receiver, uint256 amount);
 
     /**
-    * @dev Emitted with `amount` when `depositer` deposits sUSD.
-    */
+     * @dev Emitted with `amount` when `depositer` deposits sUSD.
+     */
     event Deposit(address depositer, uint256 amount);
 
     /**
-    * @dev Emitted when a position is successfully opened.
-    */
+     * @dev Emitted when a position is successfully opened.
+     */
     event PositionOpen(
         address trader,
-        uint margin,
-        uint size,
+        uint256 margin,
+        uint256 size,
         IFuturesMarket market,
         Pozition position
     );
 
     /**
-    * @dev Emitted when a position is successfully opened.
-    */
-    event PositionClose(
-        address trader,
-        IFuturesMarket market,
-        Pozition position
-    );
+     * @dev Emitted when a position is successfully opened.
+     */
+    event PositionClose(address trader, IFuturesMarket market, Pozition position);
 
     /// State Variables ///
 
-    mapping (address => uint) public depositsByWalletAddress;
+    mapping(address => uint256) public depositsByWalletAddress;
 
     /// Constructor ///
 
-    constructor (IAddressResolver _addressResolver, PozitionFactory _positionFactory) {
+    constructor(IAddressResolver _addressResolver, PozitionFactory _positionFactory) {
         addressResolver = _addressResolver;
         positionFactory = _positionFactory;
 
@@ -95,7 +91,7 @@ contract PozitionManager is ReentrancyGuard {
         markets[1] = "FuturesMarketETH";
         markets[2] = "FuturesMarketLINK";
 
-        for (uint i; i < markets.length; i++) {
+        for (uint256 i; i < markets.length; i++) {
             bytes32 name = markets[i];
             address market = addressResolver.getAddress(name);
             supportedMarkets[name] = IFuturesMarket(market);
@@ -106,11 +102,14 @@ contract PozitionManager is ReentrancyGuard {
     /// Mutative Functions ///
 
     /**
-    * @dev Deposit sUSD into the manager to be used as margin when opening positions.
-    */
-    function deposit(uint _amount) public {
+     * @dev Deposit sUSD into the manager to be used as margin when opening positions.
+     */
+    function deposit(uint256 _amount) public {
         require(_amount > 0, "Deposit amount is too small.");
-        require(marginToken.allowance(msg.sender, address(this)) >= _amount, "Approve sUSD token first!");
+        require(
+            marginToken.allowance(msg.sender, address(this)) >= _amount,
+            "Approve sUSD token first!"
+        );
 
         depositsByWalletAddress[msg.sender] += _amount;
         bool isSuccess = marginToken.transferFrom(msg.sender, address(this), _amount);
@@ -120,13 +119,16 @@ contract PozitionManager is ReentrancyGuard {
     }
 
     /**
-    * @dev Withdraw previously deposited sUSD from the manager. A `_receiver` is specified to signal
-    * the address which will be receiving the sUSD upon a successful withdraw.
-    */
-    function withdraw(uint _amount, address _receiver) public nonReentrant {
+     * @dev Withdraw previously deposited sUSD from the manager. A `_receiver` is specified to signal
+     * the address which will be receiving the sUSD upon a successful withdraw.
+     */
+    function withdraw(uint256 _amount, address _receiver) public nonReentrant {
         require(_amount > 0, "Withdraw amount is too small.");
         require(_receiver != address(0), "Receiver cannot be NULL.");
-        require(depositsByWalletAddress[msg.sender] >= _amount, "Withdrawing more than available.");
+        require(
+            depositsByWalletAddress[msg.sender] >= _amount,
+            "Withdrawing more than available."
+        );
 
         depositsByWalletAddress[msg.sender] -= _amount;
         bool isSuccess = marginToken.transfer(_receiver, _amount);
@@ -136,18 +138,18 @@ contract PozitionManager is ReentrancyGuard {
     }
 
     /**
-    * @dev Performs a trade on `_market` using `_amount` margin previously provided in the `FuturesPositionsManager`.
-    *
-    * This withdraws margin from within the `FuturesPositionsManager` but will revert if `_amount > deposit`. The
-    * address of the newly minted NFT position is returned upon successful trade.
-    *
-    * The `_sizeDelta` is a derived value from a combination of proportion leverage and the price of the underlying
-    * asset for the specified market. Synthetix Futures does the hard work work validation these attributes before
-    * allowing a position to be created. This function will revert if Synthetix Futures reverts.
-    */
+     * @dev Performs a trade on `_market` using `_amount` margin previously provided in the `FuturesPositionsManager`.
+     *
+     * This withdraws margin from within the `FuturesPositionsManager` but will revert if `_amount > deposit`. The
+     * address of the newly minted NFT position is returned upon successful trade.
+     *
+     * The `_sizeDelta` is a derived value from a combination of proportion leverage and the price of the underlying
+     * asset for the specified market. Synthetix Futures does the hard work work validation these attributes before
+     * allowing a position to be created. This function will revert if Synthetix Futures reverts.
+     */
     function openPosition(
-        uint _margin,
-        uint _size,
+        uint256 _margin,
+        uint256 _size,
         bytes32 _market,
         string memory _fullTokenURI
     ) public returns (Pozition position) {
@@ -158,7 +160,9 @@ contract PozitionManager is ReentrancyGuard {
         IFuturesMarket market = supportedMarkets[_market];
         require(address(market) != address(0), "Market not supported.");
 
-        position = Pozition(positionFactory.clone(msg.sender, market, _size, _margin, _fullTokenURI));
+        position = Pozition(
+            positionFactory.clone(msg.sender, market, _size, _margin, _fullTokenURI)
+        );
 
         withdraw(_margin, address(position));
 
@@ -169,22 +173,22 @@ contract PozitionManager is ReentrancyGuard {
     }
 
     /**
-    * @dev Given the `_position`, close if possible and withdraw all margin into the manager.
-    *
-    * TODO: Liquidated positions with remaining margin should ideally be deposited back into the manager
-    * to be withdrawn or used in another position in the future but that is not currently implemented.
-    */
+     * @dev Given the `_position`, close if possible and withdraw all margin into the manager.
+     *
+     * TODO: Liquidated positions with remaining margin should ideally be deposited back into the manager
+     * to be withdrawn or used in another position in the future but that is not currently implemented.
+     */
     function closePosition(Pozition _position) external {
         _position.closeAndBurn();
         emit PositionClose(msg.sender, _position.market(), _position);
     }
 
     /**
-    * @dev Utility method combining `deposit` and `openPosition`.
-    */
+     * @dev Utility method combining `deposit` and `openPosition`.
+     */
     function depositMarginAndOpenPosition(
-        uint _margin,
-        uint _size,
+        uint256 _margin,
+        uint256 _size,
         bytes32 _market,
         string memory _fullTokenURI
     ) external returns (Pozition position) {
